Add ::float cast to actors.tmdb_popularity query
The actors.tmdb_popularity column is also numeric/DECIMAL, so pg returns
it as a string. Add same ::float cast as the movie/show queries.

+++ server/scripts/scheduled-popularity-update.ts
@@ -447,7 +447,7 @@ async function updateActorPopularity(
     id: number
     tmdb_popularity: number | null
   }>(`
-    SELECT id, tmdb_popularity
+    SELECT id, tmdb_popularity::float
     FROM actors
     WHERE deathday IS NOT NULL
     ORDER BY id
